fix: added validation when updating a negative source with another negative source

+++ packages/backend/src/main/java/pt/com/deimos/vcwapi/service/CriteriaService.java
@@ -99,23 +99,23 @@ public class CriteriaService {
     SourceEntity oldSource = oldCriteria.getSource();
     SourceDTO newSourceInfo = editedInfo.getSource();
     SourceEntity newSource;
-    if (newSourceInfo == null) {
+    if (newSourceInfo != null && oldSource != null){
+      BeanUtils.copyProperties(newSourceInfo, oldSource);
+      oldSource.setUpdatedAt(LocalDateTime.now());
+      oldSource.setUpdatedBy(userId);
+    }
+    if (newSourceInfo == null && oldSource != null) {
       oldCriteria.setSource(null);
       oldSource.removeCriteria(oldCriteria);
       this.sourceRepository.delete(oldSource);
     }
-    else if (oldSource == null) {
+    else if (oldSource == null && newSourceInfo != null) {
       newSource = new SourceEntity();
       BeanUtils.copyProperties(newSourceInfo, newSource);
       newSource.setCreatedBy(userId);
       newSource.setUpdatedAt(LocalDateTime.now());
       newSource.setUpdatedBy(userId);
       oldCriteria.setSource(newSource);
-    }
-    else {
-      BeanUtils.copyProperties(newSourceInfo, oldSource);
-      oldSource.setUpdatedAt(LocalDateTime.now());
-      oldSource.setUpdatedBy(userId);
     }
 
     return this.criteriaRepository.save(oldCriteria);
